modifcaciones de tipo alimento dto y su delete funcional

+++ KiloApi/src/main/java/com/Triana/Salesinaos/KiloApi/controller/TipoAlimentoController.java
@@ -3,6 +3,7 @@ package com.Triana.Salesinaos.KiloApi.controller;
 import com.Triana.Salesinaos.KiloApi.dto.clase.ClaseDto;
 import com.Triana.Salesinaos.KiloApi.dto.tipoAlimento.TipoAlimentoDto;
 import com.Triana.Salesinaos.KiloApi.model.TipoAlimento;
+import com.Triana.Salesinaos.KiloApi.service.AportacionService;
 import com.Triana.Salesinaos.KiloApi.service.TipoAlimentoService;
 import io.swagger.v3.oas.annotations.OpenAPIDefinition;
 import io.swagger.v3.oas.annotations.Operation;
@@ -33,6 +34,7 @@ import java.util.stream.Collectors;
 public class TipoAlimentoController {
 
     private final TipoAlimentoService tipoAlimentoService;
+    private final AportacionService aportacionService;
 
 
     @Operation(summary = "Este método lista todos tipos de alimentos")
@@ -43,30 +45,16 @@ public class TipoAlimentoController {
                             array = @ArraySchema(schema = @Schema(implementation = TipoAlimento.class)),
                             examples = {@ExampleObject(
                                     value = """
-                                               [
+                                            [
                                                 {
                                                     "id": 1,
-                                                    "nombre": "Patatas"
+                                                    "nombre": "cerveza",
+                                                    "kilosDisponibles": 7.0
                                                 },
                                                 {
-                                                    "id": 2,
-                                                    "nombre": "Arroz"
-                                                },
-                                                {
-                                                    "id": 7,
-                                                    "nombre": "Banana"
-                                                },
-                                                {
-                                                    "id": 8,
-                                                    "nombre": "Pizza"
-                                                },
-                                                {
-                                                    "id": 9,
-                                                    "nombre": "Botella Cocacola"
-                                                },
-                                                {
-                                                    "id": 10,
-                                                    "nombre": "Mejillones"
+                                                    "id": 4,
+                                                    "nombre": "patatas",
+                                                    "kilosDisponibles": 0.0
                                                 }
                                             ]
                                             """
@@ -97,10 +85,11 @@ public class TipoAlimentoController {
                             array = @ArraySchema(schema = @Schema(implementation = TipoAlimentoDto.class)),
                             examples = {@ExampleObject(
                                     value = """
-                                            [
-                                                {"id": 2,
-                                                "nombre": "Mejillones en escabeche"},
-                                            ]
+                                            {
+                                                "id": 1,
+                                                "nombre": "cerveza",
+                                                "kilosDisponibles": 7.0
+                                            }
                                             """
                             )}
                     )}),
@@ -130,11 +119,11 @@ public class TipoAlimentoController {
                             array = @ArraySchema(schema = @Schema(implementation = TipoAlimento.class)),
                             examples = {@ExampleObject(
                                     value = """
-                                            [
-                                                {"id": 1, "nombre": "Pasta"},
-                                                {"id": 2, "nombre": "Aceite"},
-                                                {"id": 3, "nombre": "Arroz"}
-                                            ]
+                                            {
+                                                    "id": 1,
+                                                    "nombre": "cerveza",
+                                                    "kilosDisponibles": 0.0
+                                                }
                                             """
                             )}
                     )}),
@@ -160,10 +149,11 @@ public class TipoAlimentoController {
                             array = @ArraySchema(schema = @Schema(implementation = TipoAlimentoDto.class)),
                             examples = {@ExampleObject(
                                     value = """
-                                            [
-                                               {"id": 2,
-                                                "nombre": "Atún en manteca",
-                                            ]
+                                            {
+                                                "id": 1,
+                                                "nombre": "aceitunas",
+                                                "kilosDisponibles": 7.0
+                                            }
                                             """
                             )}
                     )}),
@@ -200,9 +190,8 @@ public class TipoAlimentoController {
     })
     @DeleteMapping("/{id}")
     public ResponseEntity<TipoAlimento> deleteTipoAlimento(@PathVariable Long id){
-        if(tipoAlimentoService.existById(id))
+        if(tipoAlimentoService.existById(id)&& aportacionService.findByTipoAlimentoId(id).isEmpty())
             tipoAlimentoService.deleteById(id);
-        //
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
